RDKB-55524,RDKB-55523,RDKB-55522,RDKB-55521 : Federate GPON, PPP, XDSL, VLAN Manager into Public repo.
Reason for change:
Needs to remove the header on the Gpon recipe. We don't put headers on metadata files in meta repositories, so that the blanket MIT license applies. (As specified in the license file) I cleared off the other recipes in the tool.

Test Procedure:
Build should passed and all WAN functionality should work.

Risks: Medium

Signed-off-by: Lakshminarayanan <lakshminarayanan.shenbagaraj2@sky.uk>